updated aur command

+++ bash/dependencies/aur.sh
@@ -1,15 +1,17 @@
 #! /bin/bash
 
-URL="${1}"
+URL="${2}"
 aur_install() {
-    cd "${HOME}"/Downloads/
-    curl "${URL}"
-    DOWNLOADED_FILE="${URL##*/}"
-    tar xvzf "${DOWNLOADED_FILE}"
-    EXTRACTED_DIR="${DOWNLOADED_FILE%%.*}"
-    cd "${EXTRACTED_DIR}"
-    makepkg -sicCfL
-    cd ..
-    rm "${DOWNLOADED_FILE}"
-    rm -rf "${EXTRACTED_DIR}"
+    if ! $PACSEARCH | grep "${1}" > /dev/null 2>&1
+        cd "${HOME}"/Downloads/
+        curl "${URL}"
+        DOWNLOADED_FILE="${URL##*/}"
+        tar xvzf "${DOWNLOADED_FILE}"
+        EXTRACTED_DIR="${DOWNLOADED_FILE%%.*}"
+        cd "${EXTRACTED_DIR}"
+        makepkg -sicCfL
+        cd ..
+        rm "${DOWNLOADED_FILE}"
+        rm -rf "${EXTRACTED_DIR}"
+    fi
 }
